requirements.txt
click==8.1.2; python_version >= "3.7"
colorama==0.4.4; python_version >= "3.7" and python_full_version < "3.0.0" and platform_system == "Windows" or platform_system == "Windows" and python_version >= "3.7" and python_full_version >= "3.5.0"
flask==2.1.1; python_version >= "3.7"
importlib-metadata==4.11.3; python_version < "3.10" and python_version >= "3.7"
itsdangerous==2.1.2; python_version >= "3.7"
jinja2==3.1.1; python_version >= "3.7"
joblib==1.1.0; python_version >= "3.7"
markupsafe==2.1.1; python_version >= "3.7"
numpy==1.22.3; python_version >= "3.8"
opencv-python==4.5.5.64; python_version >= "3.6"
pywavelets==1.3.0; python_version >= "3.7"
scikit-learn==1.0.2; python_version >= "3.7"
scipy==1.6.1; python_version >= "3.7"
threadpoolctl==3.1.0; python_version >= "3.7"
werkzeug==2.1.1; python_version >= "3.7"
zipp==3.8.0; python_version < "3.10" and python_version >= "3.7"
